move classes to own namespace change endpoint in gateway
