+++ Applications/ProductionManager/Contracts/OrderProgress/OrderProgressDetails.cs
@@ -1,7 +1,8 @@
 ﻿using HomagConnect.Base.Contracts.Enumerations;
+
 using Newtonsoft.Json;
 
-namespace HomagConnect.ProductionManager.Contracts.Orders
+namespace HomagConnect.ProductionManager.Contracts.OrderProgress
 {
     /// <summary>
     /// Represents processed part totals for one order at one workstation.
